+++ src/Trade360SDK.Common.Entities/Entities/OutrightLeague/OutrightLeagueMarketCompetitionWrapper.cs
@@ -1,0 +1,10 @@
+using System;
+using System.Collections.Generic;
+
+namespace Trade360SDK.Common.Entities.OutrightLeague
+{
+    public class OutrightLeagueMarketCompetitionWrapper<TEvent> : OutrightLeagueCompetitionWrapper<TEvent>
+    {
+        public DateTime? NextFixtureStartTime { get; set; }
+    }
+}